+++ mobile/android/android-components/components/feature/privatemode/src/test/java/mozilla/components/feature/privatemode/feature/SecureWindowFeatureTest.kt
@@ -84,7 +84,7 @@ class SecureWindowFeatureTest {
     @Test
     fun `remove flags on stop`() {
         val store = BrowserStore()
-        val feature = SecureWindowFeature(window, store)
+        val feature = SecureWindowFeature(window, store, clearFlagOnStop = true)
 
         feature.start()
         feature.stop()